Move AjaxRequest to separate module from AjaxClient
